Quickcheck the Pricing Rules
This commit removes some of the more extended manual tests in favour
of quickchecking the pricing rules. Random values are passed into the
pricing rules, and we verify that the output matches the properties
we would expect of a correct implementation.

The value here is a little questionable. These pricing rules are a
very simple implementation, and the Quickcheck functions pretty
much reimplement the entire pricing rule in order to identify the
expectations.

On the plus side, I now know how to use Quickcheck in Go.

A way to improve this commit is to focus on the real properties we
expect, rather than the exact values. Increasing the number of items
should always be more expensive by X (except going odd->even BOGOF.)
The BOGOF price should go up by Y when going even->odd, but not change
when going odd->even. The rate of increase of bulk discount prices
should change exactly at the minimum number of items for the discount.