Implement YAML representations of Pricing Rules
It took lots of repeated refactoring to get to this point. We now have
a YAML format:

    ```
    pricing_rules:
      bogof:
        rule: buy_one_get_one_free
        unit_price: 9
      bulk:
        rule: bulk_discount
        standard_price: 450
        minimum_number_of_items: 3
        discounted_price: 399
      fixed:
        rule: fixed
        unit_price: 11
    ```

This deserialises into `pricing_rules.Config`. The reflection needed
to specify the type of rule with the `rule` field took some doing, and
that is much of why that `Config` type exists.

As well as some hardcoded tests we have quickcheck generating random
pricing rules, serialising them to YAML, deserialising them, and
verifying that they are the same before and after.