Optimise Quadtree.Intersects
Until now, the intersection times for 500 snakes of lengths averaging
30 segments were about 28ms. This turns out to have been poorly
optimised: the quadtree was not pruning subtrees even when they didn't
overlap with the circle at all.

This commit prevents checking circles against subtrees they don't
overlap with. The speedup is variable because each circle is now
compared with a variable number of other circles.

Typical worst-case times are now 1-2ms. Average times are dozens of
microseconds.